image-segmentation/requirements.txt
setuptools>=41.0.0
Flask==1.1.1
gunicorn==20.0.4
numpy==1.16.4
Pillow==7.2.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu
--find-links https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/torch1.5/index.html
detectron2==0.2